+++ src/Heisenburger69/BurgerSpawners/Entities/Enderman.php
@@ -10,7 +10,8 @@ class Enderman extends Monster {
     public const NETWORK_ID = self::ENDERMAN;
 
     public $width = 0.3;
-    public $length = (float) 0.9;
+    /** @var float */
+    public $length = 0.9;
     public $height = 1.8;
 
     public function getName(): string{
